requirements.txt
fastapi
uvicorn
python-multipart
jinja2
reportlab
faiss-cpu
tiktoken
python-dotenv
google-generativeai
